Removed some uneccessary includes

+++ src/include/OtherFunctions/Factorial.h
@@ -3,9 +3,9 @@
 #define _FACTORIAL_
 
 template<typename T>
-uint64_t Factorial (T n)
+double Factorial (T n)
 {
-	uint64_t fact = 1;
+	double fact = 1;
 	T temp = n;
 	
 	while (temp > 1)
